User details UI (#50)
* Handle unit tests for ollama

* Fix AIController not finding prompts

* Add multi-turn unit test

* Add verbosity when AI responds and refactor multi-turn responses

* Fix multichat unit test

* layout for chat and quiz pages, just the file for the auth view.

* small commit, removed auth-view as mason already made actual views

* Fix ChatController ui methods & updated window size

* Updated Chat UI and ChatController UI methods

* Implemented create new chat functionality with added ChatSetupController and seperate chat-setup-ui page

* Updated edit and confirm chat name buttons, Added logout button and delete chat functionality

* Removed ToString override in Chat class

* updated quiz ui(temporarily in different .fxml until its finished), added QuizStyles.css, renamed ChatStyles.css, cleaned up importts in the views

* updated preliminary quiz ui to make interaction implementation easier

* Implement return button functionality

* Fixed chatController unit tests failing

* Add create chat functionality to main button

* Add multi-turn unit test

* User settings UI rehaul

# Conflicts:
#	src/main/resources/ai/tutor/cab302exceptionalhandlers/user-settings-view.fxml

* Update User settings functionality

* Add alerts when handling exceptions

* Refactor alerts and fix minor errors

* Add temporary stat labels in user settings

* Seperate FXML logic from CRUD logic

---------

+++ src/main/java/ai/tutor/cab302exceptionalhandlers/controller/UserSettingsController.java
@@ -3,11 +3,11 @@ package ai.tutor.cab302exceptionalhandlers.controller;
 import ai.tutor.cab302exceptionalhandlers.Utils.Utils;
 import ai.tutor.cab302exceptionalhandlers.model.*;
 import javafx.fxml.FXML;
-import javafx.scene.Node;
 import javafx.scene.control.*;
 import javafx.stage.Stage;
 
 import java.sql.SQLException;
+import java.util.Optional;
 
 public class UserSettingsController {
     private UserDAO userDAO;
@@ -19,20 +19,27 @@ public class UserSettingsController {
     @FXML private Button backButton;
     @FXML private Button deleteUserButton;
     @FXML private TextField usernameField;
-    @FXML private TextField currentPasswordField;
-    @FXML private TextField newPasswordField;
-    @FXML private TextField confirmPasswordField;
+    @FXML private PasswordField passwordField;
+    @FXML private PasswordField newPasswordField;
     @FXML private TextField studyArea;
     @FXML private ComboBox educationLevelCombo;
+    @FXML private Label usernameFeedback;
+    @FXML private Label passwordFeedback;
+    @FXML private Label newPasswordFeedback;
 
     // User stats widgets
     @FXML private Label quizzesTakenLabel;
     @FXML private Label averageScoreLabel;
 
+    private boolean usernameChanged;
+    private boolean passwordChanged;
+
     public UserSettingsController(SQLiteConnection connection, User user) throws SQLException {
         db = connection;
         currentUser = user;
         userDAO = new UserDAO(db);
+        this.usernameChanged = false;
+        this.passwordChanged = false;
     }
 
     private Stage getStage() {
@@ -59,27 +66,36 @@ public class UserSettingsController {
     }
 
 
-    // TODO: Add separate buttons for change username and password
     private void setupSaveButton(){
         saveButton.setOnAction(actionEvent -> {
-            try {
-                String username = usernameField.getText();
-                String currentPassword = currentPasswordField.getText();
-                String newPassword = newPasswordField.getText();
-                String confirmPassword = confirmPasswordField.getText();
+            usernameFeedback.setText(null);
+            passwordFeedback.setText(null);
+            newPasswordFeedback.setText(null);
 
-                if (!newPassword.equals(confirmPassword)) {
-                    throw new SecurityException("Passwords do not match");
-                }
+            String currentPasswordText = passwordField.getText();
+            String newPasswordText = newPasswordField.getText();
+            String usernameText = usernameField.getText();
 
-                if (Utils.validateNullOrEmpty(currentPassword) && Utils.validateNullOrEmpty(newPassword)) {
-                    updateUsername(username);
-                } else {
-                    updatePassword(newPassword, currentPassword);
-                }
-                Utils.showInfoAlert("User details updated");
-            } catch (Exception e) {
-                Utils.showErrorAlert(e.getMessage());
+            usernameChanged = false;
+            passwordChanged = false;
+
+            handleUsernameInput(usernameText);
+            handlePasswordInput(newPasswordText, currentPasswordText);
+
+            String successMessage = null;
+
+            if (usernameChanged && passwordChanged) {
+                successMessage = "Username & Password updated";
+            } else if (usernameChanged) {
+                successMessage = "Username updated";
+            } else if (passwordChanged) {
+                successMessage = "Password updated";
+            }
+
+            if (successMessage != null) {
+                Utils.showInfoAlert(successMessage);
+                passwordField.setText(null);
+                newPasswordField.setText(null);
             }
         });
     }
@@ -87,28 +103,37 @@ public class UserSettingsController {
     private void setupDeleteButton(){
         deleteUserButton.setOnAction(actionEvent -> {
             try {
-
-                //TODO: Should probably make a popup asking user to confirm
-                userDAO.deleteUser(currentUser);
-
+                Optional<ButtonType> result = Utils.showConfirmAlert("Are you sure you want to logout of your account?");
+                if (result.isPresent()) {
+                    ButtonType buttonClicked = result.get();
+                    if (buttonClicked == ButtonType.OK) {
+                        userDAO.deleteUser(currentUser);
+                    }
+                }
             } catch (SQLException e) {
-                e.printStackTrace();
+                Utils.showErrorAlert("Failed to logout");
             }
 
             try {
                 Utils.loadView("login", new LoginController(db), getStage());
             } catch (Exception e ) {
-                Utils.showErrorAlert("Error Logging Out: " + e);
+                Utils.showErrorAlert("Error Logging Out");
             }
         });
     }
 
     private void setupLogoutButton(){
         logoutButton.setOnAction(actionEvent -> {
-            try {
-                Utils.loadView("login", new LoginController(db), getStage());
-            } catch (Exception e ) {
-                Utils.showErrorAlert("Error Logging Out: " + e);
+            Optional<ButtonType> result = Utils.showConfirmAlert("Are you sure you want to logout?");
+            if (result.isPresent()) {
+                ButtonType buttonClicked = result.get();
+                if (buttonClicked == ButtonType.OK) {
+                    try {
+                        Utils.loadView("login", new LoginController(db), getStage());
+                    } catch (Exception e ) {
+                        Utils.showErrorAlert("Error Logging Out");
+                    }
+                }
             }
         });
     }
@@ -118,9 +143,40 @@ public class UserSettingsController {
             try {
                 Utils.loadView("chat", new ChatController(db, currentUser), getStage());
             } catch (Exception e ) {
-                Utils.showErrorAlert("Error Returning To Chat: " + e);
+                Utils.showErrorAlert("Error Returning To Chat");
             }
         });
+    }
+
+    private void handlePasswordInput(String newPassword, String currentPasswordText) {
+        try {
+            updatePassword(newPassword, currentPasswordText);
+        } catch (IllegalArgumentException e) {
+            if (e.getMessage().contains("null or empty")) {
+                newPasswordFeedback.setText("Enter new password");
+            } else if (e.getMessage().contains("invalid")) {
+                passwordFeedback.setText("Invalid password, try again");
+            }
+        } catch (SecurityException e) {
+            passwordFeedback.setText("Incorrect password");
+        }
+        catch (SQLException e) {
+            Utils.showErrorAlert("Failed to change password");
+        }
+    }
+
+    private void handleUsernameInput(String username) {
+        try {
+            updateUsername(username);
+        } catch (IllegalArgumentException e) {
+            if (e.getMessage().contains("already taken")) {
+                usernameFeedback.setText("Username already taken");
+            } else if (e.getMessage().contains("invalid")) {
+                usernameFeedback.setText("Invalid username, try again");
+            }
+        } catch (SQLException e) {
+            Utils.showErrorAlert("Failed to change username");
+        }
     }
 
 
@@ -131,31 +187,38 @@ public class UserSettingsController {
      */
 
     public void updateUsername(String newUsername) throws IllegalArgumentException, SQLException {
-        if (!User.validUsername(newUsername)) {
+        if (!User.validUsername(newUsername)){
             throw new IllegalArgumentException("Username is invalid");
         }
 
-        if (!newUsername.equals(currentUser.getUsername())) {
+        if (!currentUser.getUsername().equals(newUsername)) {
             User existingUser = userDAO.getUser(newUsername);
             if (existingUser != null) {
                 throw new IllegalArgumentException("Username is already taken");
             }
+            currentUser.setUsername(newUsername);
+            userDAO.updateUser(currentUser);
+            usernameChanged = true;
         }
-
-        currentUser.setUsername(newUsername);
-        userDAO.updateUser(currentUser);
     }
 
     public void updatePassword(String newPassword, String currentPassword) throws IllegalArgumentException, SecurityException, SQLException {
-        if (!User.validPassword(newPassword)) {
-            throw new IllegalArgumentException("Password is invalid");
-        }
-        if (!currentUser.verifyPassword(currentPassword != null ? currentPassword : "")) {
-            throw new SecurityException("Incorrect Password");
-        }
+        if (!Utils.validateNullOrEmpty(newPassword) || !Utils.validateNullOrEmpty(currentPassword)) {
+            if (!Utils.validateNullOrEmpty(currentPassword) && Utils.validateNullOrEmpty(newPassword) ) {
+                throw new IllegalArgumentException("New password is null or empty");
+            }
 
-        String hashedNewPassword = User.hashPassword(newPassword);
-        currentUser.setPasswordHash(hashedNewPassword);
-        userDAO.updateUser(currentUser);
+            if (!User.validPassword(currentPassword)) {
+                throw new IllegalArgumentException("Password is invalid");
+            }
+
+            if (!currentUser.verifyPassword(currentPassword)){
+                throw new SecurityException("Incorrect Password");
+            }
+            String newPasswordHash = User.hashPassword(newPassword);
+            currentUser.setPasswordHash(newPasswordHash);
+            userDAO.updateUser(currentUser);
+            passwordChanged = true;
+        }
     }
 }
